refactor: move tak detection to new file
Also: remove unnecessary regex dependency
