+++ Cargo.toml
@@ -24,5 +24,4 @@ serde = { version = "1.0.214", features = ["derive"] }
 thiserror = "1.0.68"
 
 [dev-dependencies]
-regex = { version = "1.11.1" }
 serde_json = { version = "1.0.132" }
