+++ src/tak/mod.rs
@@ -1,73 +1,4 @@
-/// Support for detecting common variants of CoT messages used by TAK, and parsing them into
-/// strongly-typed structs.
-use crate::{
-    detail::{parse, CotUnparsedDetail},
-    Error,
-};
-
-/// An enum of expected message types.
-#[derive(Copy, Clone, Debug, PartialEq)]
-pub enum TakCotType {
-    GeoFence,
-    Marker,
-    RangeBearing,
-    Route,
-    Shape,
-    Other,
-}
-
-/// Result of parsing a TAK CoT message and attempting to detect a message type.
-/// We don't actually parse the `<detail>` section, since it is so dynamic in practice.
-/// Thus, callers are responsible for further parsing of that section, for example using
-/// [`quick_xml::Reader`].
-pub struct TakCotMessage {
-    pub cot_type: TakCotType,
-    pub cot_msg: CotUnparsedDetail,
-}
-
-/// Use a heuristic to detect the type of TAK XML CoT message contained in supplied text.
-/// Warning: This is based on known example messages from ATAK repository. Please do extended
-/// integration testing and file issues for any bugs you find.
-pub fn detect_tak_cot_type(input: &str) -> Result<TakCotMessage, Error> {
-    let cot_msg = parse(input)?;
-    // List of tuples of <string to search for> -> <implied message type if found>
-    // Order matters. Separated into base event `type` values and detail section values to search
-    // for.
-    let type_tokens = [
-        ("u-r-b-", TakCotType::RangeBearing),
-        ("u-rb-", TakCotType::RangeBearing),
-        ("b-m-r", TakCotType::Route),
-        ("u-d-", TakCotType::Shape),
-    ];
-    let detail_tokens = [
-        ("__geofence", TakCotType::GeoFence),
-        ("usericon", TakCotType::Marker),
-    ];
-    // First, search detail section for clues
-    for line in &cot_msg.detail {
-        for (search, msg_type) in detail_tokens.iter() {
-            if line.contains(search) {
-                return Ok(TakCotMessage {
-                    cot_type: *msg_type,
-                    cot_msg,
-                });
-            }
-        }
-    }
-    // Next check the base event type
-    for (search, msg_type) in type_tokens.iter() {
-        if cot_msg.cot_type.contains(search) {
-            return Ok(TakCotMessage {
-                cot_type: *msg_type,
-                cot_msg,
-            });
-        }
-    }
-    Ok(TakCotMessage {
-        cot_type: TakCotType::Other,
-        cot_msg,
-    })
-}
+pub mod detect;
 
 #[cfg(test)]
 mod test {
@@ -83,38 +14,9 @@ mod test {
     fn test_tak_cot_examples() {
         let examples = get_xml_examples().unwrap();
         for res in examples {
-            let (filename, cot_xml) = res.unwrap();
-            let cot = parse(&cot_xml).unwrap();
-            println!("{filename}\n\t{:?}", cot);
+            let (_filename, cot_xml) = res.unwrap();
+            let _cot = parse(&cot_xml).unwrap();
         }
-    }
-
-    #[test]
-    fn test_tak_cot_detect() {
-        let examples = get_xml_examples().unwrap();
-        for res in examples {
-            let (filename, cot_xml) = res.unwrap();
-            let cot = crate::tak::detect_tak_cot_type(&cot_xml).unwrap();
-            if filename.starts_with("geo-fence") {
-                assert_type(&filename, cot.cot_type, TakCotType::GeoFence);
-            } else if filename.starts_with("marker-") {
-                assert_type(&filename, cot.cot_type, TakCotType::Marker);
-            } else if filename.starts_with("range-bearing") {
-                assert_type(&filename, cot.cot_type, TakCotType::RangeBearing);
-            } else if filename.starts_with("route") {
-                assert_type(&filename, cot.cot_type, TakCotType::Route);
-            } else if filename.starts_with("shape-") {
-                assert_type(&filename, cot.cot_type, TakCotType::Shape);
-            }
-        }
-    }
-
-    fn assert_type(filename: &str, actual: TakCotType, expected: TakCotType) {
-        assert_eq!(
-            actual, expected,
-            "{}: expected {:?}, got {:?}",
-            filename, expected, actual
-        );
     }
 
     /// You can use serde_json::Value for storing dynamic XML data, except for the issue with
@@ -137,20 +39,11 @@ mod test {
         println!("after json round trip:\n\t{}", xml1);
     }
 
-    fn get_xml_examples() -> Result<CotExamples, Error> {
+    pub fn get_xml_examples() -> Result<CotExamples, Error> {
         let examples_path = format!("{}/src/tak/examples", env!("CARGO_MANIFEST_DIR"));
         let examples = CotExamples::new(examples_path).unwrap();
         assert!(examples.len() > 0);
         Ok(examples)
-    }
-    use regex::Regex;
-
-    use super::TakCotType;
-
-    #[allow(dead_code)]
-    fn trim_whitespace(xml: &str) -> String {
-        let re = Regex::new(r">\s+<").unwrap();
-        re.replace_all(xml, "><").to_string()
     }
 
     // Test helper to iterate over all example messages
